+++ src/model/guild/system_channel.rs
@@ -4,7 +4,7 @@ bitflags! {
     /// [Discord docs](https://docs.discord.com/developers/resources/guild#guild-object-system-channel-flags).
     #[cfg_attr(feature = "typesize", derive(typesize::derive::TypeSize))]
     #[derive(Copy, Clone, Default, Debug, Eq, Hash, PartialEq)]
-    pub struct SystemChannelFlags: u64 {
+    pub struct SystemChannelFlags: u8 {
         /// Suppress member join notifications.
         const SUPPRESS_JOIN_NOTIFICATIONS = 1 << 0;
         /// Suppress server boost notifications.
